eigen性能优化

+++ wheeled_bipedal_controllers/include/wheeled_bipedal_controllers/ins_task.h
@@ -8,25 +8,25 @@
 
 typedef struct
 {
-    float q[4]; // 四元数估计值
+    double q[4]; // 四元数估计值
 
-    float Gyro[3];          // 角速度
-    float Accel[3];         // 加速度
-    float MotionAccel_b[3]; // 机体坐标加速度
-    float MotionAccel_n[3]; // 绝对系加速度
+    double Gyro[3];          // 角速度
+    double Accel[3];         // 加速度
+    double MotionAccel_b[3]; // 机体坐标加速度
+    double MotionAccel_n[3]; // 绝对系加速度
 
-    float AccelLPF; // 加速度低通滤波系数
+    double AccelLPF; // 加速度低通滤波系数
 
     // 加速度在绝对系的向量表示
-    float xn[3];
-    float yn[3];
-    float zn[3];
+    double xn[3];
+    double yn[3];
+    double zn[3];
 
     // 位姿
-    float Roll;
-    float Pitch;
-    float Yaw;
-    float YawTotalAngle;
+    double Roll;
+    double Pitch;
+    double Yaw;
+    double YawTotalAngle;
 } INS_t;
 
 extern INS_t INS;   // imu
@@ -34,7 +34,7 @@ extern INS_t INS;   // imu
 void INS_Init();
 void INS_Task(double aX, double aY, double aZ, double gX, double gY, double gZ, double dt);
 
-void BodyFrameToEarthFrame(const float *vecBF, float *vecEF, float *q);
-void EarthFrameToBodyFrame(const float *vecEF, float *vecBF, float *q);
+void BodyFrameToEarthFrame(const double *vecBF, double *vecEF, double *q);
+void EarthFrameToBodyFrame(const double *vecEF, double *vecBF, double *q);
 
 #endif
